+++ src/modules/file.ts
@@ -11,6 +11,13 @@ export function InsertWeb(path: string) {
     return path.splice(index != -1 ? index : path.length, 0, web);
 }
 
+export function RemoveWeb(path: string, sep: string = '/') {
+    const web = "WEB";
+    let newPath = path.replace(sep + web, '');
+    newPath = path == newPath ? path.replace(web + sep, '') : newPath;
+    return newPath;
+}
+
 
 /**
  * Converts local file path to remote path using user config
@@ -94,7 +101,7 @@ export async function GetAllFilesInDir(startPath: string): Promise<string[]> {
     return files;
 }
 
-export interface SimpleFolder{
+export interface SimpleFolder {
     path: string,
     files: string[],
     folders: SimpleFolder[],
@@ -104,14 +111,14 @@ export async function GetAllFilesInDirTree(startPath: string): Promise<SimpleFol
     if (!await exists(startPath)) {
         return null;
     }
-    const folder: SimpleFolder = { path: startPath, files: [], folders: []};
+    const folder: SimpleFolder = { path: startPath, files: [], folders: [] };
     const directory = await readdir(startPath);
     const promises = [];
     for (const file of directory) {
         const name = path.join(startPath, file);
         const stat = await lstat(name);
         if (stat.isDirectory()) {
-            promises.push(GetAllFilesInDirTree(name).then((cFolder)=>{
+            promises.push(GetAllFilesInDirTree(name).then((cFolder) => {
                 return folder.folders.push(cFolder);
             }));
         }